+++ include/SI/detail/unit_cast.h
@@ -1,28 +1,9 @@
 #pragma once
 
+#include "detail.h"
 #include <ratio>
 
 namespace SI::detail {
-
-// forward declaration
-template <char _symbol, typename _exponent, typename _type, typename _ratio>
-struct unit_t;
-
-/// helper template to check if a type is a unit_t (false for all other
-/// types)
-template <typename _unit> struct is_unit_t : std::false_type {};
-
-/// template specialisation to check if a type is a unit_t (true if unit_t)
-template <char _symbol, typename _exponent, typename _ratio, typename _type>
-struct is_unit_t<const unit_t<_symbol, _exponent, _type, _ratio>>
-    : std::true_type {};
-
-/// non-const specialisation of check above
-template <char _symbol, typename _exponent, typename _ratio, typename _type>
-struct is_unit_t<unit_t<_symbol, _exponent, _type, _ratio>> : std::true_type {};
-
-template <typename _type>
-inline constexpr bool is_unit_t_v = is_unit_t<_type>::value;
 
 /// function to cast between two units of the same type
 template <typename _target_type, typename _rhs_T>
@@ -42,4 +23,19 @@ constexpr auto unit_cast(const _rhs_T &rhs) {
   return _target_type(
       ((rhs.value() * conversion_ratio::num) / conversion_ratio::den));
 }
+
+template <typename _unit_lhs, typename _unit_rhs>
+struct unit_with_common_ratio {
+  static_assert(is_unit_t_v<_unit_lhs>, "only supported for SI::unit_t");
+  static_assert(is_unit_t_v<_unit_rhs>, "only supported for SI::unit_t");
+  static_assert(std::is_convertible<typename _unit_lhs::internal_type,
+                                    typename _unit_rhs::internal_type>::value);
+  static_assert(_unit_lhs::symbol::value == _unit_rhs::symbol::value);
+  using type =
+      unit_t<_unit_lhs::symbol::value, typename _unit_lhs::exponent,
+             typename _unit_lhs::internal_type,
+             typename detail::ratio_gcd<typename _unit_lhs::ratio,
+                                        typename _unit_rhs::ratio>::ratio>;
+};
+
 } // namespace SI::detail